correction lien

+++ Makefile
@@ -765,7 +765,7 @@ pcFixeConf : install
 
 
 install-Data:
-	rmdir ~/Documents
+	[ ! -L ~/Documents ] && rmdir ~/Documents
 	@ln -snf /DATA/Documents ~/Documents
 	@ln -snf /DATA/Obsidian ~/Obsidian
 	@ln -snf /DATA/Projets ~/Projets
@@ -774,7 +774,7 @@ install-myTools: sanity-check
 	yes|sudo pacman -Syyu
 	yes|sudo pacman -Syy
 	yes|sudo pacman -S --noconfirm --needed xorg-xev
-	yes|sudo pacman -S --noconfirm --needed drawio-desktop keepassxc obsidian calibre thunderbird darktable simple-scan syncthing freecad inkscape remmina
+	yes|sudo pacman -S --noconfirm --needed drawio-desktop keepassxc obsidian calibre thunderbird darktable simple-scan syncthing freecad inkscape remmina gimp thunar
 	yes|sudo pacman -S --noconfirm --needed texlive-latex  texlive-latexextra texlive-langfrench texlive-fontsextra texlive-latexrecommended texlive pandoc-cli
 	[ ! -d ~/Documents/ThemeSSR ] && git clone --depth=1 https://gitlab.com/willo22/ThemeSSR.git ~/Documents/ThemeSSR/
 	sudo chmod +x ~/Documents/ThemeSSR/scripts/*.sh
